feat(redis): redis cluster support for queue and locking
- Enables use of Redis Cluster for Orca's queue and distributed lock
service, when SQL is used for the execution repository. Running Orca
entirely on Redis Cluster remains unsupported.
