+++ gradle.properties
@@ -2,6 +2,6 @@
 fiatVersion=1.1.0
 enablePublishing=false
 spinnakerGradleVersion=6.5.0
-korkVersion=5.8.5
-keikoVersion=2.10.1
+korkVersion=5.8.8
+keikoVersion=2.12.0
 org.gradle.parallel=true
