+++ orca-redis/orca-redis.gradle
@@ -2,7 +2,9 @@ apply from: "$rootDir/gradle/kotlin.gradle"
 apply from: "$rootDir/gradle/spock.gradle"
 
 dependencies {
-  api("redis.clients:jedis")
+  api("redis.clients:jedis:2.10.2") {
+    force = true
+  }
   api("com.netflix.spinnaker.kork:kork-jedis")
 
   implementation(project(":orca-core"))
